# fix deleteat! and subset! performance