Explanation in https://discourse.julialang.org/t/learning-to-benchmark-and-find-the-best-function-to-select-a-subset-of-a-dataframe/91704/12.

# Benchmarks

## This PR
```
julia> using Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> df = DataFrame(rand(10^6, 100), :auto);

julia> df.id = rand(1:100, 10^6);

julia> inds = rand(Bool, 10^6);

julia> x = copy(df); @time deleteat!(x, inds);
  0.084989 seconds (29.05 k allocations: 1.471 MiB, 11.83% compilation time)

julia> x = copy(df); @time deleteat!(x, inds);
  0.074356 seconds (305 allocations: 4.766 KiB)

julia> x = copy(df); @time subset!(x, :x1 => Returns(inds));
  0.091337 seconds (485 allocations: 262.422 KiB)

julia> x = copy(df); @time subset!(x, :x1 => Returns(inds));
  0.148983 seconds (485 allocations: 262.422 KiB)

julia> x = groupby(copy(df), :id); @time subset!(x, :x1 => x -> rand(Bool, length(x)));     
  0.207012 seconds (1.15 M allocations: 63.651 MiB, 31.11% compilation time)

julia> x = groupby(copy(df), :id); @time subset!(x, :x1 => x -> rand(Bool, length(x)));     
  0.216079 seconds (1.15 M allocations: 63.632 MiB, 30.41% compilation time)
```

## 1.4.4 release
```
julia> using Random

julia> Random.seed!(1234)
TaskLocalRNG()

julia> df = DataFrame(rand(10^6, 100), :auto);

julia> df.id = rand(1:100, 10^6);

julia> inds = rand(Bool, 10^6);

julia> x = copy(df); @time deleteat!(x, inds);
  0.300327 seconds (307 allocations: 3.817 MiB)

julia> x = copy(df); @time deleteat!(x, inds);
  0.303013 seconds (307 allocations: 3.817 MiB)

julia> x = copy(df); @time subset!(x, :x1 => Returns(inds));
  0.307057 seconds (483 allocations: 4.074 MiB)

julia> x = copy(df); @time subset!(x, :x1 => Returns(inds));
  0.300665 seconds (483 allocations: 4.074 MiB)

julia> x = groupby(copy(df), :id); @time subset!(x, :x1 => x -> rand(Bool, length(x)));
  0.437468 seconds (1.15 M allocations: 67.427 MiB, 15.85% compilation time)

julia> x = groupby(copy(df), :id); @time subset!(x, :x1 => x -> rand(Bool, length(x)));
  0.417350 seconds (1.15 M allocations: 67.425 MiB, 15.70% compilation time)
```