@nalimilan 2022-12-17
Interesting. Have you checked with more columns and with a lower percentage of dropped rows? I would expect the `findall` approach to be less slow (and maybe faster) in these cases.