@bkamins 2022-12-17
> I would expect the `findall` approach to be less slow (and maybe faster) in these cases.

You are right! 🧠

The threshold value I assessed empirically is less than 5% observations when it is better (it probably also depends on number of columns, but I wanted to have something relatively simple). I have proposed an adaptive algorithm switching between two approaches as needed.