@nalimilan 2022-12-17
Cool. Can you check when there are many columns? That's a use case that we care about.