@bkamins 2022-12-17
Here is an example: 100 columns, 10^6 rows. Tested with 5.5% rows to drop (so a bit above 5% threshold).

Setup:
```
df = DataFrame(rand(10^6, 100), :auto)
```

This PR:
```
julia> t = 0.055;

julia> Random.seed!(1234);

julia> idx = rand(10^6) .< t;

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.095975 seconds (303 allocations: 4.734 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.096654 seconds (303 allocations: 4.734 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.094744 seconds (303 allocations: 4.734 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.094374 seconds (303 allocations: 4.734 KiB)
```

Current release:
```
julia> t = 0.055;

julia> Random.seed!(1234);

julia> idx = rand(10^6) .< t;

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.092951 seconds (304 allocations: 432.203 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.087335 seconds (304 allocations: 432.203 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.091471 seconds (304 allocations: 432.203 KiB)

julia> dfc = copy(df); @time deleteat!(dfc, idx);
  0.092145 seconds (304 allocations: 432.203 KiB)
```

I did some more tests on even wider tables and it seems that a more precise threshold is 6% on my laptop, so I changed it to that value.